Use Table when have personal ID

+++ app/src/main/java/com/sharpdroid/registroelettronico/Databases/Entities/EventInfo.kt
@@ -4,7 +4,7 @@ import com.orm.SugarRecord
 
 data class EventInfo(
         var remote: Boolean,
-        val eventId: Int,
+        val eventId: Long,
         var completed: Boolean,
         var archived: Boolean
 ) : SugarRecord()
